v0.3.18

+++ ranx/data_structures/qrels.py
@@ -297,7 +297,9 @@ class Qrels(object):
         assert (
             df[doc_id_col].dtype == "O"
         ), "DataFrame Document IDs column dtype must be `object` (string)"
-        assert df[score_col].dtype == int, "DataFrame scores column dtype must be `int`"
+        assert (
+            df[score_col].dtype == np.int64
+        ), "DataFrame scores column dtype must be `int`"
 
         qrels_dict = (
             df.groupby(q_id_col)[[doc_id_col, score_col]]
